Refactor code a little
