+++ src/renin/ui/performancePanel.glsl
