+++ src/style.css
@@ -9,5 +9,5 @@
 
 @font-face {
   font-family: 'Barlow';
-  src: url(./renin/fonts/Barlow-Regular.ttf) format('truetype');
+  src: url(./renin/ui/fonts/Barlow-Regular.ttf) format('truetype');
 }
